Adds JavaFX GUI following tutorial part 4
